Closet component now gets the correct data from the server based on which button the user clicks. Next step: displaying that data.
