+++ frontend/outfitcreator/src/Components/Closet_Page/closet.css
@@ -1,0 +1,7 @@
+.closet-text {}
+
+.closet-menu {}
+
+.closet-button {
+    margin: 10px 5px;
+}
